Templates: do not configure mode tcp based on ip-protocol
It looks like ip-protocol is to determine id protocol is TCP or
UDP. It is not suited to distinguish tcp from http proxies.

Use /Common/http profile to setup mode http in fronted.

This patch also setup mode http by default for backends. A comment
instructs to change it to tcp for tcp proxies.

Signed-off-by: Sébastien Gross <seb•ɑƬ•chezwam•ɖɵʈ•org>